- moved conversion algorithm to separate module
